# Conflicts:
#	E-StockMarketApplication/src/main/java/com/example/EStockMarketApplication/DTOs/CompanyResponseDTO.java

+++ E-StockMarketApplication/src/main/java/com/example/EStockMarketApplication/Repository/StockPriceRepository.java
@@ -10,6 +10,6 @@ import java.lang.Long;
 import java.util.Optional;
 
 @Repository
-public interface StockPriceRepository extends JpaRepository<StockPrice,Integer> {
+public interface StockPriceRepository extends JpaRepository<StockPrice,Long> {
 
 }
